% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestPartitions.R
\name{bestPartitions}
\alias{bestPartitions}
\title{Best number of cluster partitions.}
\usage{
bestPartitions(data, krange = 2:10, cluster.method = "UnsupRF",
  dist.method = "euclidean", parallel = FALSE, mc.cores = 2,
  control = NULL, sep.measure = "SW", ensemble = TRUE)
}
\arguments{
\item{data}{data matrix to cluster}

\item{krange}{integer vector. Numbers of clusters to try}

\item{dist.method}{Use \code{\link{RFdist}} or R's base \code{\link{dist}} method to compute dissimilarity 
matrix. Default is "UnsupRF".}

\item{parallel}{run in parallel ? Currently uses \code{\link{mcapply}} which works only on linux operating systems}

\item{mc.cores}{number of CPU cores}

\item{control}{list with control parameters: 
\itemize{
\item ntree number of trees for \code{\link{RFdist}}
\item no.rep number of repetitions or forests for \code{\link{RFdist}}
\item neighbSize  number of nearest neighbors to compute connectivity validation measure. See the function 
 \code{\link{clusterConnectivity}}
\item method \code{hclust} method 
\item RF.parallel  type of parallization for \code{\link{RFdist}}
\item compact.measure  one of five compactness measure. See \code{\link{clusterCompactness}}
\item combined: if ensemble = FALSE, should the seperation measures be combined?  
}}

\item{ensemble}{take the ensemble of the seperation measures ?}
}
\value{
clustering for data
}
\description{
Get "best" number of clustering partitions determined by \code{\link{successiveOmitClusterValidation}} for 
k = 2, ..., K using cluster connectivity. Specifically, this function runs \code{\link{successiveOmitClusterValidation}}  
for each cluster number in krange and selects the optimal number of cluster by looking at the 
clustering with the best separation or compactnest
}
\details{
This function is experimental .. work in progress to validate the procedure
}
