% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCompactness.R
\name{clusterCompactness}
\alias{clusterCompactness}
\title{cluster compactness measures}
\usage{
clusterCompactness(distance = NULL, clusters, Data = NULL,
  method = "euclidean", ensemble = TRUE, compact.measure = "SW", ...)
}
\arguments{
\item{distance}{dissimilarity matrix}

\item{clusters}{existing cluster partitions to determine compactness}

\item{ensemble}{whether to find the optimal compactness measure by majority rule}

\item{compact.measure}{if ensemble=FALSE, provide the compactness measure to return. Options are 
\itemize{
\item DM  cluster diabmeter 
\item SW  Cluster Silhouette width
\item WD  Within cluster distance 
\item gap  Cluster gap
}}

\item{\dots}{further arguments passed to or from other methods.}

\item{data}{data matrix from which cluster partitions were derived. Note that one of  \code{\link{distance}} 
or \code{\link{data}} must be provided}
}
\value{
the numeric compactness measure
}
\description{
Calculate 5 within cluster compactness or homogenerity measures
}
