% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{clusterConnectivity}
\alias{clusterConnectivity}
\title{cluster connectivity measure}
\usage{
clusterConnectivity(distance = NULL, clusters, Data = NULL,
  neighbSize = 10, method = "euclidean")
}
\description{
Calculates the connectivity validation measure for data points or dissimilarity matrix and
 cluster partitioning. Currently uses R's \code{link{dist}} method.
}
