% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSeparation.R
\name{clusterSeparation}
\alias{clusterSeparation}
\title{cluster Separation measures}
\usage{
clusterSeparation(distance = NULL, Data = NULL, clusters,
  method = "euclidean", DB = TRUE, ...)
}
\description{
Calculate 5 between cluster separation measures  #
}
