% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStability.R
\name{clusterStability}
\alias{clusterStability}
\alias{Stability}
\title{Cluster Stability. 
\deqn{J = \frac{n_{11}}{n_{11} + n_{10} + n_{01}}}.}
\usage{
Stability(c1, c2)
}
\arguments{
\item{c1}{a vector of \code{n} clustering labels}

\item{c2}{a vector of \code{n} clustering labels}
}
\value{
the Jaccard coefficient for the two sets of cluster labels (See
Details.)
}
\description{
Computes the cluster stability measure defined in Famili et.al based on the 
idea of cluster immovability on partition. Cluster immovability is the rate at which 
the content of a cluster remains unchanged during the clustering process.
}
\examples{
\dontrun{
# We generate K = 3 labels for each of n = 10 observations and compute the
# Jaccard similarity coefficient between the two clusterings.
set.seed(42)
K <- 3
n <- 10
labels1 <- sample.int(K, n, replace = TRUE)
labels2 <- sample.int(K, n, replace = TRUE)
jaccard_indep(labels1, labels2)

# Here, we cluster the iris data set with the K-means and
# hierarchical algorithms using the true number of clusters, K = 3.
# Then, we compute the Jaccard similarity coefficient between the two
# clusterings.
iris_kmeans <- kmeans(iris[, -5], centers = 3)$cluster
iris_hclust <- cutree(hclust(dist(iris[, -5])), k = 3)
jaccard_indep(iris_kmeans, iris_hclust)
}
}
\references{
Famili, A. Fazel, Ganming Liu, and Ziying Liu. "Evaluation and optimization of clustering in 
gene expression data analysis." Bioinformatics 20.10 (2004): 1535-1545.
}
