% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictStrength.R
\name{clvpredictStrength}
\alias{clvpredictStrength}
\alias{clvpredictStrength.default}
\alias{plot.clvpredictStrength}
\alias{print.clvpredictStrength}
\title{cluster validation by prediction strength}
\usage{
clvpredictStrength(dat, ...)

\method{clvpredictStrength}{default}(dat, dist.mat, method = "ward.D2",
  nBoots = 20, krange = 2:5, balanced = FALSE, parallel = FALSE,
  mc.cores = 2, OOB = TRUE, seed = 12345, ...)

\method{plot}{clvpredictStrength}(x, perf.measure = "Sensitivity")

\method{print}{clvpredictStrength}(x, ...)
}
\arguments{
\item{dat}{a data matrix}

\item{\dots}{further arguments passed to or from other methods.}

\item{dist.mat}{dissimilarity matrix obtained from the data matrix "dat"}

\item{method}{\code{\link{hclust}} clustering method}

\item{nBoots}{number of bootstraps}

\item{krange}{integer vector. Numbers of clusters which are to be tried}

\item{parallel}{run in parallel ?}

\item{mc.cores}{number of CPU cores}

\item{OOB}{use out-of-bag from bootstrap as test data?}

\item{seed}{random seed}

\item{balance}{perform balance bootstrap ?}
}
\description{
Calculate the mediods of a clustering by finding the point that has the minimum average/sum 
distance to all other points in the cluster. Proximity between data points can be 
provided by \code{RFdist} or using the Gower's general similarity coefficient.
}
\details{
This function is experimental .. more details coming
}
