% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediod.R
\name{mediod}
\alias{mediod}
\alias{mediod.default}
\alias{mix.dist}
\alias{print.mediod}
\title{mediods of a clustering procedure}
\usage{
mediod(x, ...)

\method{mediod}{default}(x, clusters, fun = "sum", weights, ...)

mix.dist(data, weights)

\method{print}{mediod}(x, ...)
}
\arguments{
\item{x}{a dissimilarity matrix, or a data frame/ matrix}

\item{\dots}{further arguments passed to or from other methods.}

\item{clusters}{clustering}

\item{fun}{character name of the function to determine minimum 
distance between points in a cluster. Should be either \code{mean}, 
\code{median} or \code{sum} (default)}

\item{weights}{ptional vector of weights for variables in data. See \code{growdis} in the 
\code{FD} package}

\item{data}{data frame/ matrix}
}
\value{
data matrix of cluster mediods and the corresponding row indicies of 
the mediods in the original data
}
\description{
Calculate the mediods of a clustering by finding the point that has the minimum average/sum 
distance to all other points in the cluster. Proximity between data points can be 
provided by \code{RFdist} or using the Gower's general similarity coefficient.
}
\details{
\code{mediod} is the main function to compute the mediod, while 
\code{mix.dist} computes the distance between obersavations based on veriables 
of mix-types: binary, categorical, and continuous using the Gower's general similarity 
coefficient.
}
\references{
Gower, John C. "A general coefficient of similarity and some of its properties." Biometrics (1971): 857-871.
}
