% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nselectboot.R
\name{nselectboot}
\alias{nselectboot}
\title{Selection of optimal number of clusters using bootstrap.
This is a minor modification of the  nselectboot function in the package fpc, so you can use package::nselectboot
to run the perferred version if both packaes are loaded. See \code{\link[fpc]{nselectboot}} for details.}
\usage{
nselectboot(data, B = 50, distances = inherits(data, "dist"),
  clustermethod = NULL, classification = "averagedist",
  krange = 2:10, count = FALSE, nnk = 1, ...)
}
\arguments{
\item{data}{data or dissimilarity matrix}

\item{B}{number of boostrap}

\item{distances}{(logical) is data a dissimilarity matrix ?}

\item{clustermethod}{clustering method, see \code{\link[fpc]{nselectboot}}}

\item{classification}{see \code{\link[fpc]{nselectboot}}}

\item{krange}{integer vector; numbers of clusters to be tried}

\item{count}{logical. If TRUE, numbers of clusters and bootstrap runs are printed.}

\item{nnk}{number of closest neighbors for classification}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Selection of optimal number of clusters using bootstrap.
This is a minor modification of the  nselectboot function in the package fpc, so you can use package::nselectboot
to run the perferred version if both packaes are loaded. See \code{\link[fpc]{nselectboot}} for details.
}
