% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mediod}
\alias{predict.mediod}
\title{predict cluster memberships for new data.}
\usage{
\method{predict}{mediod}(object, org.data = NULL, newdata, ...)
}
\arguments{
\item{object}{object of class \code{\link{mediod}}. This requires \code{mediod} to 
contain the data matrix of the mediod.}

\item{org.data}{(optional) the original data used to generate the mediods.}

\item{newdata}{A new input data frame.}

\item{...}{Further arguments passed to mix.dist.}
}
\value{
clustering for new data
}
\description{
Predict cluster memberships for new data by locating points closest to the mediods of a
a clustering procedure. Distance between mediod and  newdata is based on the Gower's general similarity 
coefficient, which applies to mix-type data.
}
\examples{
\dontrun{
set.seed(12345)
data(iris)
dat <- iris[, -5]
RF.dist <- RFdist(data=dat, ntree = 10, no.rep=20, syn.type = "permute", 
               importance=FALSE)
# run hclust 
HCmod <- hclust(d=RF.dist$RFdist, method="ward.D2", members= NULL)
clusters <- cutree(HCmod, 3)   
# get mediods of clustering 
 med <- mediod(x = RF.dist$RFdist, clusters=clusters, fun = "sum")
 print(med)
# predict clusters 
pred.clusters <- predict(object= med, org.data = dat, newdata = dat) 
table(clusters, pred.clusters)
}
}
