% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFdissimilarity.R
\name{RFdist}
\alias{RFdist}
\alias{RFdist.default}
\alias{print.RFdist}
\alias{plot.RFdist}
\title{Unsupervised Random Forest}
\usage{
RFdist(data, ...)

\method{RFdist}{default}(data, mtry = floor(sqrt(ncol(data))), ntree,
  no.rep, syn.type = "emperical", importance = FALSE, nodesize = 1,
  parallel = c("forests", "trees", "no"), ...)

\method{print}{RFdist}(x, ...)

\method{plot}{RFdist}(x, ...)
}
\arguments{
\item{data}{data.frame or matrix}

\item{\dots}{further arguments passed to \code{\link[randomForest]{randomForest}}.}

\item{mtry}{mtry in \code{\link[randomForest]{randomForest}}}

\item{ntree, }{number of trees}

\item{no.rep}{number of repetitions or forests}

\item{syn.type}{type of synthetic data generator: "emperical" generate samples from the 
emperical distribution of the original data while "permute" takes a permutation 
of each dimension. emerical is just sample with 
replacement while permute is without replacement}

\item{importance}{(logical) compute variable importance  ?}

\item{nodesize}{node size in \code{randomForest}}

\item{parallel}{character vector specifying the type of parallel run: 'forests' - run a total of 
\code{no.rep} RF in parallel, 
'trees' - run RF in parallel over ntrees and combine the results, 'no' - serial computation.}

\item{x}{object of class \code{\link{RFdist}}}
}
\value{
A list with elements:  
\enumerate{
\item RFdist: RF proximity converted to a distance object
\item err: error rate
\item UnsupRFvarimp: Unsupervised RF variable importance 
\item proxConver: a matrix containing three convergence meausres 
                  \enumerate{                        
                  \item  Max.prox = max( abs( aveprox(N)- aveprox(N-1)))  
                  \item MSE.prox = mean( (aveprox(N)- aveprox(N-1))^2)   
	             \item Mean =  mean(aveprox(N))
                  where N is number of forests (no.rep).
                   }         
}
}
\description{
Compute dissimilarity matrix between observations by training a 
randomForest (RF) classifier to 
descriminate between the 'original' data and a synthetic version. 
The original data is labeled as "True.Data" 
while the synthetic data is labeled ``Synthetic.Data". The random forest p
roximity matrix between observations in the 
original data are then extracted, converted to distance, and returned.  
The synthetic data is generated by taking a random sample from each 
dimension of the true data, with 
or without replacement.
}
\details{
Methods 
\enumerate{
\item \code{print} : print OOB error and convergence summary  
\item \code{plot} : plots the convergence of the RF proximities given by the MSE 
over number of forest \code{no.rep}
}
}
\examples{
\dontrun{
set.seed(12345)
data(iris)
dat <- iris[, -5]
RF.dist <- RFdist(data=dat, ntree = 10, no.rep=20, syn.type = "permute", 
               importance=TRUE, parallel = "no")
# 
print(RF.dist, digits = 3)
#  
plot(RF.dist)
# plot variable importance 
UnsupRFvarImpPlot(RF.dist, sort=TRUE)
}
}
\references{
Tao Shi and Steve Horvath (2006) Unsupervised Learning with Random Forest Predictors. 
Journal of Computational and Graphical Statistics. Volume 15, Number 1, March 2006, pp. 118-138(21)
}
