% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterValidation.R
\name{successiveOmitClusterValidation}
\alias{successiveOmitClusterValidation}
\title{Cluster Validation by Successive Omission}
\usage{
successiveOmitClusterValidation(data, k = 2, vars,
  cluster.method = "UnsupRF", dist.method = "euclidean",
  control = NULL, ensemble = TRUE, sep.measure = "SW")
}
\description{
Validate a clustering scheme by succerssively omitting the most compact/connected cluster member and  
 re-cluster the remaining data points, and continue the process all over.   
The hypothesis is that, if the data can be clustered into k optimal cluster, then after
 dropping the most compact/connected cluster member, the remaining data points can best be clustered 
into k-1 clusters and so on, untill k = 2. By trying different k's, the successive omission algorithm may
identify the best k clusters for the data.
}
\details{
This function is experimental and not meant to be used directly by the user
Work in progress to validate the procedure
}
