% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupRFfpc.R
\name{UnsupRF}
\alias{UnsupRF}
\alias{UnsupRF.default}
\alias{print.UnsupRF}
\title{Unsupervised random forest clustering with fpc.}
\usage{
UnsupRF(data, ...)

\method{UnsupRF}{default}(data, RFdist, B = 10,
  clustermethod = pamkCBI, classification = "centroid", krange = 2:5,
  kopt = 2, run.boot = FALSE, fun = "sum", ...)

\method{print}{UnsupRF}(x, ...)
}
\arguments{
\item{data}{data.frame or matrix}

\item{\dots}{further arguments passed to or from other methods.}

\item{RFdist}{RF distance matrix computed from \code{\link{RFdist}}.}

\item{B}{number of bootstraps}

\item{clustermethod}{clustering method, options are \code{pamkCBI}, or \code{claraCBI}, or \code{hclustCBI}. 
  Not to sure about \code{hclustCBI} see the \code{fpc} package. \code{pamkCBI} is 
  recommended for RF dissimilarity matrix, but we have found standard 
\code{hclust} in base R works well with  Ward's minimum variance creterion}

\item{classification}{type of prediction for finding optimal number of clusters 
see \code{\link[fpc]{nselectboot}}.}

\item{krange}{integer vector; numbers of clusters to be tried}

\item{kopt}{user provided optimal number of clusters}

\item{run.boot}{(logical) run bootstrap cluster-wise stability ?}

\item{fun}{function to determine mediods, should be \code{mean}, \code{median}, 
or \code{sum}. See \code{\link{mediod}}}

\item{x}{object of class \code{\link{UnsupRF}}}
}
\value{
A list with elements:  
\enumerate{
\item cluster.model: The cluster model  
\item cluster: cluster memberships 
\item kopt: optimal number of clusters  
\item mediods: a mediod object            
}
}
\description{
Unsupervised random forest clustering. A Random forest (RF) classifier is trained to predict 
the data labeled as class ``True.Data" and a synthetic data labeled as class ``Synthetic.Data". 
The synthetic data is generated by taking a random sample from each dimension of the true data, with 
or without replacement (see \code{\link{RFdist}}). The dissimilarity matrix from \code{\link{RFdist}} is 
then passed to the algorithms in the "flexible point clustering" 
\code{fpc} package for clustering and selection of optimal number of clusters through the bootstrap 
cluster-wise stability method.
}
\examples{
\dontrun{
set.seed(12345)
data(iris)
dat <- iris[, -5]
RF.dist <- RFdist(data=dat, ntree = 10, no.rep=20, syn.type = "permute", 
               importance=TRUE, oob.prox=TRUE, proxConver=FALSE)
#            
Clus.res <- UnsupRF(data = dat, RFdist=RF.dist$RFdist, 
             B =  5, clustermethod=pamkCBI, classification="centroid", 
             krange= 2:4, kopt=2, run.boot = TRUE)
 print(Clus.res)            
clusters <- Clus.res$clusters 
kopt <- Clus.res$kopt # optimal number of clusters 
}
}
