% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnsupRvarImpPlot.R
\name{UnsupRFvarImpPlot}
\alias{UnsupRFvarImpPlot}
\title{Unsupervised random forest variable importance}
\usage{
UnsupRFvarImpPlot(x, sort = TRUE, top = min(30, nrow(x$UnsupRFvarimp)),
  plot = TRUE)
}
\arguments{
\item{x}{object of class \code{\link{RFdist}}}

\item{sort, }{sort variable importance by variables in the orginal data.}

\item{top}{number of top variables to plot.}

\item{plot}{(logical) plot or simply return variable importance as a matrix ?}
}
\value{
matrix if plot = FALSE
}
\description{
Extract variable importance from \code{\link{RFdist}} if importance = TRUE  and plot for both the 
"Synthetic and "True" data
}
